Split header in .h and .cpp
