+++ rlog/rlog.h
@@ -22,7 +22,6 @@
 
 #pragma once
 
-#include <dlfcn.h>
 namespace rlog {
 
   // API functions -------------------------------------------------------------
@@ -49,136 +48,5 @@ namespace rlog {
   int localVersion = 1;
 
   // END API functions ---------------------------------------------------------
-
-namespace {
-    const char *domain = "";
-    const char *category = "";
-
-    // Static function pointers
-    void (*log_mark_) (const char*, const char*, const char*, const char*) = NULL;
-    void (*log_rangePush_) (const char*, const char*, const char*, const char*) = NULL;
-    void (*log_rangePop_) () = NULL;
-    void (*log_registerActiveCallback_) (void (*cb)()) = NULL;
-    bool (*log_isActive_) () = NULL;
-    const char* (*log_getProperty_) (const char*, const char*, const char*) = NULL;
-
-    // Static function pointers - ROCTX
-    void (*roctx_mark_) (const char* message) = NULL;
-    void (*roctx_rangePush_) (const char* message) = NULL;
-    void (*roctx_rangePop_) () = NULL;
-} // namespace
-
-
-
-// Load library and look up symbols
-void init() {
-#if 1
-    void (*dl) = dlopen("librlog.so", RTLD_LAZY);
-    if (dl) {
-        log_mark_ = (void (*)(const char*, const char*, const char*, const char*)) dlsym(dl, "rlog_mark");
-        log_rangePush_ = (void (*)(const char*, const char*, const char*, const char*)) dlsym(dl, "rlog_rangePush");
-        log_rangePop_ = (void (*)()) dlsym(dl, "rlog_rangePop");
-        log_registerActiveCallback_ = (void(*)(void (*cb)())) dlsym(dl, "rlog_registerActiveCallback");
-        log_isActive_ = (bool (*)()) dlsym(dl, "rlog_isActive");
-        log_getProperty_ = (const char*(*)(const char*, const char*, const char*))  dlsym(dl, "rlog_getProperty");
-    }
-#endif
-#if 0
-    void (*dltx) = dlopen("libroctx64.so", RTLD_LAZY);
-    if (dltx) {
-        roctx_mark_ = (void (*)(const char*)) dlsym(dltx, "roctxMarkA");
-        roctx_rangePush_ = (void (*)(const char*)) dlsym(dltx, "roctxRangePushA");
-        roctx_rangePop_ = (void (*)()) dlsym(dltx, "roctxRangePop");
-    }
-#endif
-}
-
-void mark(const char *domain, const char *category, const char *apiname, const char *args)
-{
-    if (log_mark_)
-        log_mark_(domain, category, apiname, args);
-    if (roctx_mark_) {
-        char buff[4096];
-        snprintf(buff, 4096, "%s : %s : api = %s | %s", domain, category, apiname, args);
-        roctx_mark_(buff); 
-    }
-}
-
-void mark(const char *category, const char *apiname, const char *args)
-{
-    mark(domain, category, apiname, args);
-}
-
-void mark(const char *apiname, const char *args)
-{
-    mark(domain, category, apiname, args);
-}
-
-void rangePush(const char *domain, const char *category, const char *apiname, const char *args)
-{
-//fprintf(stderr, "rangePush %p\n", log_rangePush_);
-    if (log_rangePush_)
-       log_rangePush_(domain, category, apiname, args);
-    if (roctx_rangePush_) {
-        char buff[4096];
-        snprintf(buff, 4096, "%s : %s : api = %s | %s", domain, category, apiname, args);
-        roctx_rangePush_(buff);
-    }
-}
-
-void rangePush(const char *category, const char *apiname, const char *args)
-{
-    rangePush(domain, category, apiname, args);
-}
-
-void rangePush(const char *apiname, const char *args)
-{
-    rangePush(domain, category, apiname, args);
-}
-
-void rangePop()
-{
-    if (log_rangePop_)
-        log_rangePop_();
-    if (roctx_rangePop_)
-        roctx_rangePop_();
-}
-
-
-int registerActiveCallback(void (*cb)())
-{
-    if (log_registerActiveCallback_) {
-        log_registerActiveCallback_(cb);
-        return 0;
-    }
-    else
-        return -1;
-}
-
-bool isActive()
-{
-    if (log_isActive_)
-        return log_isActive_();
-    else
-        return false;
-}
-
-void setDefaultDomain(const char* ddomain)
-{
-    domain = ddomain;
-}
-
-void setDefaultCategory(const char* dcat)
-{
-    category = dcat;
-}
-
-// FIXME: lifetime
-const char *getProperty(const char *domain, const char *property, const char *defaultValue)
-{
-    if (log_getProperty_)
-        return log_getProperty_(domain, property, defaultValue);
-    return defaultValue;
-}
 
 } // namespace rlog
